% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm2df.R
\name{gadm2df}
\alias{gadm2df}
\title{Import ESRI shapefiles into R data frame}
\usage{
gadm2df(sp_df)
}
\description{
Import shapefile and convert into lat/lon coordinates to plot in R or Tableau
}
\details{
View available projections with 'projInfo(type = 'proj')'
View available datum with 'projInfo(type = 'datum')'
View available ellipsoids with 'projInfo(type = 'ellps')'
}
\examples{

}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

