% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_crs.R
\name{lookup_crs}
\alias{lookup_crs}
\title{Lookup CRS country-level projection string based on country name or code}
\usage{
lookup_crs(country)
}
\arguments{
\item{country}{country name or id.}
}
\description{
Matches *exactly* (though case insensitive) to find the CRS PROJ4 string for a
suitable equal area projection for a given country.
}
