% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot a basic map}
\usage{
plot_map(df, fill_var, stroke_size = 0.15, stroke_colour = "white",
  file_name = NA, plot_width = 5, plot_height = 5)
}
\description{
Given a data frame containing lat/long, make a basic map
}
\examples{

}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}
