% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sp_join.R
\name{plot_sp_join}
\alias{plot_sp_join}
\title{Create leaflet map to check spatial join}
\usage{
plot_sp_join(df, shp, base_map = providers$Esri.WorldGrayCanvas,
  radius = 100)
}
\description{
after performing a spatial join using spatial_join,
creates an interactive leaflet map to check that the polygon/point matching looks correct
}
