% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2df.R
\name{sp2df}
\alias{sp2df}
\title{Convert a spatial data frame into a series of lat/lon coordinates}
\usage{
sp2df(sp_df)
}
\description{
Convert a spatial data frame into a series of lat/lon coordinates
}
\examples{

}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

