% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shp.R
\name{read_shp}
\alias{read_shp}
\title{Import ESRI shapefiles into R}
\usage{
read_shp(workingDir = getwd(), layerName)
}
\arguments{
\item{workingDir}{string containing the name of the folder containing the shapefile}

\item{layerName}{string containing the name of the shapefile, e.g. 'district_boundary' for 'district_boundary.shp'}
}
\description{
Import shapefile into an R spatial object
}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

